LayoutTests/fast/events/constructors/message-event-constructor-data-identity-expected.txt

Tests that MessageEvent.data returns the value it was initialized to.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* JS Object
Using constructor
PASS messageEvent.data === expectedData is true
PASS messageEvent.data === messageEvent.data is true
Using deprecated initMessageEvent()
PASS messageEvent.data === expectedData is true
PASS messageEvent.data === messageEvent.data is true

* String
Using constructor
PASS messageEvent.data === expectedData is true
PASS messageEvent.data === messageEvent.data is true
Using deprecated initMessageEvent()
PASS messageEvent.data === expectedData is true
PASS messageEvent.data === messageEvent.data is true

* Blob
Using constructor
PASS messageEvent.data === expectedData is true
PASS messageEvent.data === messageEvent.data is true
Using deprecated initMessageEvent()
PASS messageEvent.data === expectedData is true
PASS messageEvent.data === messageEvent.data is true

* ArrayBuffer
Using constructor
PASS messageEvent.data === expectedData is true
PASS messageEvent.data === messageEvent.data is true
Using deprecated initMessageEvent()
PASS messageEvent.data === expectedData is true
PASS messageEvent.data === messageEvent.data is true

PASS successfullyParsed is true

TEST COMPLETE
